refactor(frontend/dom): reorganize folder structure for modularity
