+++ projekt/public/js/dom/dashboard/selectors.js
