+++ projekt/public/js/dom/shared/elements.js
@@ -1,0 +1,7 @@
+//	Erzeugt ein HTML-Element mit Textinhalt
+export function createTextElement(tagName, text){
+	const el = document.createElement(tagName);
+	el.textContent = text;
+	
+	return el;
+}
